requirements.txt
qovery-client==0.1.0
Flask==1.0.2
psycopg2-binary==2.8.4
